+++ src/plugins/dev-mode/check-schema.ts
@@ -11,6 +11,7 @@ import { getPrimaryFieldOfPrimaryKey, getSchemaByObjectPath } from '../../rx-sch
 import type {
     CompositePrimaryKey,
     JsonSchema,
+    JsonSchemaTypes,
     RxJsonSchema,
     TopLevelProperty
 } from '../../types';
@@ -340,6 +341,93 @@ export function checkSchema(jsonSchema: RxJsonSchema<any>) {
                     }
                 }
             }
+
+            /**
+             * To be able to craft custom indexable string with compound fields,
+             * we need to know the maximum fieldlength of the fields values
+             * when they are transformed to strings.
+             * Therefore we need to enforce some properties inside of the schema.
+             */
+            const indexAsArray = isMaybeReadonlyArray(index) ? index : [index];
+            indexAsArray.forEach(fieldName => {
+                const schemaPart = getSchemaByObjectPath(
+                    jsonSchema,
+                    fieldName
+                );
+
+
+                const type: JsonSchemaTypes = schemaPart.type as any;
+                switch (type) {
+                    case 'string':
+                        const maxLength = schemaPart.maxLength;
+                        if (!maxLength) {
+                            throw newRxError('SC34', {
+                                index,
+                                field: fieldName,
+                                schema: jsonSchema
+                            });
+                        }
+                        break;
+                    case 'number':
+                    case 'integer':
+                        const multipleOf = schemaPart.multipleOf;
+                        if (!multipleOf) {
+                            throw newRxError('SC35', {
+                                index,
+                                field: fieldName,
+                                schema: jsonSchema
+                            });
+                        }
+                        const maximum = schemaPart.maximum;
+                        const minimum = schemaPart.minimum;
+                        if (
+                            typeof maximum === 'undefined' ||
+                            typeof minimum === 'undefined'
+                        ) {
+                            throw newRxError('SC37', {
+                                index,
+                                field: fieldName,
+                                schema: jsonSchema
+                            });
+                        }
+                        break;
+                    case 'boolean':
+                        /**
+                         * If a boolean field is used as an index,
+                         * it must be required.
+                         */
+                        let parentPath = '';
+                        let lastPathPart = fieldName;
+                        if (fieldName.includes('.')) {
+                            const partParts = fieldName.split('.');
+                            lastPathPart = partParts.pop();
+                            parentPath = partParts.join('.');
+                        }
+                        const parentSchemaPart = getSchemaByObjectPath(
+                            jsonSchema,
+                            parentPath
+                        );
+                        if (
+                            !parentSchemaPart.required ||
+                            !parentSchemaPart.required.includes(lastPathPart)
+                        ) {
+                            throw newRxError('SC38', {
+                                index,
+                                field: fieldName,
+                                schema: jsonSchema
+                            });
+                        }
+                        break;
+
+                    default:
+                        throw newRxError('SC36', {
+                            fieldName,
+                            type: schemaPart.type as any,
+                            schema: jsonSchema,
+                        });
+                }
+            });
+
         });
     }